Add copy button to table result output + move `renderResultAsJson` to shared utils
