+++ shared/common/components/resultGrid/index.tsx
@@ -1,6 +1,9 @@
+import {useEffect} from "react";
 import {observer} from "mobx-react-lite";
 
 import cn from "@edgedb/common/utils/classNames";
+import {renderResultAsJson} from "@edgedb/common/utils/renderJsonResult";
+import {CopyButton} from "../../newui/copyButton";
 
 import {GridHeader, ResultGridState, RowHeight} from "./state";
 
@@ -11,14 +14,13 @@ import {
   GridHeaders,
   HeaderResizeHandle,
 } from "../dataGrid";
+import {calculateInitialColWidths} from "../dataGrid/utils";
 import {renderCellValue} from "../dataGrid/renderUtils";
 
 import gridStyles from "../dataGrid/dataGrid.module.scss";
 import inspectorStyles from "@edgedb/inspector/inspector.module.scss";
 
 import styles from "./resultGrid.module.scss";
-import {useEffect} from "react";
-import {calculateInitialColWidths} from "../dataGrid/utils";
 
 export {createResultGridState, ResultGridState} from "./state";
 
@@ -28,6 +30,7 @@ export interface ResultGridProps extends Omit<DataGridProps, "state"> {
 }
 
 export const ResultGrid = observer(function ResultGrid({
+  className,
   state,
   bottomPadding,
   ...props
@@ -51,9 +54,18 @@ export const ResultGrid = observer(function ResultGrid({
   }, [state.grid.gridContainerSize.width]);
 
   return (
-    <DataGrid state={state.grid} {...props}>
+    <DataGrid
+      className={cn(styles.resultGrid, className)}
+      state={state.grid}
+      {...props}
+    >
       <ResultGridHeaders state={state} />
       <ResultGridContent state={state} bottomPadding={bottomPadding} />
+
+      <CopyButton
+        className={styles.copyButton}
+        content={() => renderResultAsJson(state.data, state.codec, null)}
+      />
     </DataGrid>
   );
 });
